moved things around a bit and cleaned up a bit
