+++ service_event/app/no/uio/musit/microservice/event/dao/EnvRequirementDao.scala
@@ -1,9 +1,9 @@
 package no.uio.musit.microservice.event.dao
 
-import no.uio.musit.microservice.event.domain.Dto
+import no.uio.musit.microservice.event.domain.{Dto, EnvRequirementDto}
 import no.uio.musit.microservices.common.domain.MusitInternalErrorException
 import play.api.Play
-import play.api.db.slick.{ DatabaseConfigProvider, HasDatabaseConfig }
+import play.api.db.slick.{DatabaseConfigProvider, HasDatabaseConfig}
 import play.api.libs.json.Json
 import slick.driver.JdbcProfile
 
@@ -13,7 +13,7 @@ import scala.concurrent.Future
  * Created by ellenjo on 6/30/16.
  */
 
-object EnvRequirementDAO extends HasDatabaseConfig[JdbcProfile] {
+object EnvRequirementDao extends HasDatabaseConfig[JdbcProfile] {
 
   import driver.api._
 
@@ -21,16 +21,6 @@ object EnvRequirementDAO extends HasDatabaseConfig[JdbcProfile] {
 
   private val EnvRequirementTable = TableQuery[EnvRequirementTable]
 
-  case class EnvRequirementDto(id: Option[Long], temperature: Option[Int], tempInterval: Option[Int], airHumidity: Option[Int],
-      airHumInterval: Option[Int], hypoxicAir: Option[Int], hypoxicInterval: Option[Int],
-      cleaning: Option[String], light: Option[String]) extends Dto {
-
-    //def validateInterval = if(tempInterval.get > 0) tempInterval else throw new MusitInternalErrorException("Can't have negative intervals")
-  }
-
-  object EnvRequirementDto {
-    implicit val format = Json.format[EnvRequirementDto]
-  }
 
   def insertAction(event: EnvRequirementDto): DBIO[Int] =
     EnvRequirementTable += event
