+++ service_event/app/no/uio/musit/microservice/event/service/EnvRequirementService.scala
@@ -1,7 +1,6 @@
 package no.uio.musit.microservice.event.service
 
-import no.uio.musit.microservice.event.dao.EnvRequirementDAO
-import no.uio.musit.microservice.event.dao.EnvRequirementDAO.EnvRequirementDto
+import no.uio.musit.microservice.event.dao.EnvRequirementDao
 import no.uio.musit.microservice.event.dao.EventDao.BaseEventDto
 import no.uio.musit.microservice.event.domain._
 import play.api.libs.json.{ JsObject, JsResult, Json }
@@ -19,11 +18,11 @@ object EnvRequirementService extends MultipleTablesMultipleDtos {
 
   def createEventInMemory(baseProps: BaseEventProps, customDto: Dto): Event = new EnvRequirement(baseProps, customDto.asInstanceOf[EnvRequirementDto])
 
-  def getCustomDtoFromDatabase(id: Long, baseEventProps: BaseEventProps): Future[Option[Dto]] = EnvRequirementDAO.getEnvRequirement(id)
+  def getCustomDtoFromDatabase(id: Long, baseEventProps: BaseEventProps): Future[Option[Dto]] = EnvRequirementDao.getEnvRequirement(id)
 
   def createInsertCustomDtoAction(id: Long, event: Event) = {
     val specificEvent = event.asInstanceOf[EnvRequirement]
-    EnvRequirementDAO.insertAction(specificEvent.envReqDto.copy(id = Some(id)))
+    EnvRequirementDao.insertAction(specificEvent.envReqDto.copy(id = Some(id)))
   }
 
   def validateCustomDto(jsObject: JsObject): JsResult[Dto] = jsObject.validate[EnvRequirementDto]
