+++ service_event/app/no/uio/musit/microservice/event/service/ObservationService.scala
@@ -1,6 +1,6 @@
 package no.uio.musit.microservice.event.service
 
-import no.uio.musit.microservice.event.domain.{ BaseEventProps, Event, SingleTableSingleDto }
+import no.uio.musit.microservice.event.domain.{ BaseEventProps, Event }
 
 class Observation(baseProps: BaseEventProps) extends Event(baseProps) {
   val subObservations = this.getAllSubEventsAs[Observation] //TEMP!   //Can be subtyped to SpecificObservation when that has been created
